@@ -85,12 +85,12 @@ class RecipeEditState extends State<RecipeEdit> {
       'ratio1': [
         _editedRecipe.ratio1,
         Validators.min(0.0),
-        Validators.max(100.0),
+        Validators.max(5.0),
       ],
       'ratio2': [
         _editedRecipe.ratio2,
         Validators.min(0.0),
-        Validators.max(100.0),
+        Validators.max(20.0),
       ],
       'tempSteam': [_editedRecipe.tempSteam],
       'tempWater': [_editedRecipe.tempWater],
@@ -124,7 +124,9 @@ class RecipeEditState extends State<RecipeEdit> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: widget.title == null ? Text(S.of(context).screenRecipeEditTitle) : Text(widget.title!),
+        title: widget.title == null
+            ? Text(S.of(context).screenRecipeEditTitle)
+            : Text(widget.title!),
         actions: <Widget>[
           ElevatedButton(
             child: const Text(
@@ -203,10 +205,12 @@ class RecipeEditState extends State<RecipeEdit> {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    Text("Profile", style: Theme.of(context).textTheme.labelMedium),
+                    Text("Profile",
+                        style: Theme.of(context).textTheme.labelMedium),
                     ProfileSelect(
                       onChanged: (p0) {
                         _editedRecipe.profileId = p0.id;
+                        _editedRecipe.profileName = p0.title;
                       },
                     ),
                   ],
@@ -220,7 +224,8 @@ class RecipeEditState extends State<RecipeEdit> {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    Text("Beans", style: Theme.of(context).textTheme.labelMedium),
+                    Text("Beans",
+                        style: Theme.of(context).textTheme.labelMedium),
                     BeanSelect(
                       onChanged: (coffeeId) {
                         _editedRecipe.coffee.targetId = coffeeId;
@@ -228,7 +233,8 @@ class RecipeEditState extends State<RecipeEdit> {
                           if (coffeeId == 0) {
                             Navigator.push(
                               context,
-                              MaterialPageRoute(builder: (context) => const CoffeeEdit(0)),
+                              MaterialPageRoute(
+                                  builder: (context) => const CoffeeEdit(0)),
                             );
                           } else {
                             coffeeService.setSelectedCoffee(coffeeId);
@@ -256,9 +262,11 @@ class RecipeEditState extends State<RecipeEdit> {
                     width: 200,
                     child: ReactiveTextField<double>(
                       formControlName: 'grinderSettings',
-                      keyboardType: const TextInputType.numberWithOptions(signed: true, decimal: true),
+                      keyboardType: const TextInputType.numberWithOptions(
+                          signed: true, decimal: true),
                       decoration: InputDecoration(
-                        labelText: S.of(context).screenRecipeEditGrinderSettings,
+                        labelText:
+                            S.of(context).screenRecipeEditGrinderSettings,
                       ),
                     ),
                   ),
@@ -286,7 +294,8 @@ class RecipeEditState extends State<RecipeEdit> {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(S.of(context).screenRecipeEditDosingAndWeights, style: Theme.of(context).textTheme.labelMedium),
+              Text(S.of(context).screenRecipeEditDosingAndWeights,
+                  style: Theme.of(context).textTheme.labelMedium),
               Row(
                 children: [
                   SizedBox(
@@ -303,8 +312,10 @@ class RecipeEditState extends State<RecipeEdit> {
                       },
                       showErrors: (control) => control.invalid,
                       validationMessages: {
-                        ValidationMessage.max: (error) => 'A value greater than ${(error as Map)['max']} is not accepted',
-                        ValidationMessage.min: (error) => 'A value lower than ${(error as Map)['min']} is not accepted',
+                        ValidationMessage.max: (error) =>
+                            'A value greater than ${(error as Map)['max']} is not accepted',
+                        ValidationMessage.min: (error) =>
+                            'A value lower than ${(error as Map)['min']} is not accepted',
                       },
                     ),
                   ),
@@ -325,8 +336,10 @@ class RecipeEditState extends State<RecipeEdit> {
                       },
                       showErrors: (control) => control.invalid,
                       validationMessages: {
-                        ValidationMessage.max: (error) => 'A value greater than ${(error as Map)['max']} is not accepted',
-                        ValidationMessage.min: (error) => 'A value lower than ${(error as Map)['min']} is not accepted',
+                        ValidationMessage.max: (error) =>
+                            'A value greater than ${(error as Map)['max']} is not accepted',
+                        ValidationMessage.min: (error) =>
+                            'A value lower than ${(error as Map)['min']} is not accepted',
                       },
                     ),
                   ),
@@ -347,8 +360,10 @@ class RecipeEditState extends State<RecipeEdit> {
                       },
                       showErrors: (control) => control.invalid,
                       validationMessages: {
-                        ValidationMessage.max: (error) => 'A value greater than ${(error as Map)['max']} is not accepted',
-                        ValidationMessage.min: (error) => 'A value lower than ${(error as Map)['min']} is not accepted',
+                        ValidationMessage.max: (error) =>
+                            'A value greater than ${(error as Map)['max']} is not accepted',
+                        ValidationMessage.min: (error) =>
+                            'A value lower than ${(error as Map)['min']} is not accepted',
                       },
                     ),
                   ),
@@ -365,8 +380,10 @@ class RecipeEditState extends State<RecipeEdit> {
                       ),
                       showErrors: (control) => control.invalid,
                       validationMessages: {
-                        ValidationMessage.max: (error) => 'A value greater than ${(error as Map)['max']} is not accepted',
-                        ValidationMessage.min: (error) => 'A value lower than ${(error as Map)['min']} is not accepted',
+                        ValidationMessage.max: (error) =>
+                            'A value greater than ${(error as Map)['max']} is not accepted',
+                        ValidationMessage.min: (error) =>
+                            'A value lower than ${(error as Map)['min']} is not accepted',
                       },
                     ),
                   ),
@@ -382,21 +399,26 @@ class RecipeEditState extends State<RecipeEdit> {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(S.of(context).screenRecipeEditAdjustments, style: Theme.of(context).textTheme.labelMedium),
+              Text(S.of(context).screenRecipeEditAdjustments,
+                  style: Theme.of(context).textTheme.labelMedium),
               Row(
                 children: [
                   SizedBox(
                     width: 200,
                     child: ReactiveTextField<double>(
                       formControlName: 'adjustedTemp',
-                      keyboardType: const TextInputType.numberWithOptions(signed: true, decimal: true),
+                      keyboardType: const TextInputType.numberWithOptions(
+                          signed: true, decimal: true),
                       decoration: InputDecoration(
-                        labelText: S.of(context).screenRecipeEditTemperatureCorrection,
+                        labelText:
+                            S.of(context).screenRecipeEditTemperatureCorrection,
                       ),
                       showErrors: (control) => control.invalid,
                       validationMessages: {
-                        ValidationMessage.max: (error) => 'A value greater than ${(error as Map)['max']} is not accepted',
-                        ValidationMessage.min: (error) => 'A value lower than ${(error as Map)['min']} is not accepted',
+                        ValidationMessage.max: (error) =>
+                            'A value greater than ${(error as Map)['max']} is not accepted',
+                        ValidationMessage.min: (error) =>
+                            'A value lower than ${(error as Map)['min']} is not accepted',
                       },
                     ),
                   ),
@@ -406,7 +428,10 @@ class RecipeEditState extends State<RecipeEdit> {
               Row(
                 mainAxisAlignment: MainAxisAlignment.start,
                 children: [
-                  Expanded(child: Text(S.of(context).screenRecipeEditDisableStoponweightSelectThisForPourOverWhereYouDo)),
+                  Expanded(
+                      child: Text(S
+                          .of(context)
+                          .screenRecipeEditDisableStoponweightSelectThisForPourOverWhereYouDo)),
                   SizedBox(
                     width: 120,
                     child: ReactiveSwitch(
@@ -426,10 +451,13 @@ class RecipeEditState extends State<RecipeEdit> {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(S.of(context).screenRecipeEditMilkAndWater, style: Theme.of(context).textTheme.labelMedium),
+              Text(S.of(context).screenRecipeEditMilkAndWater,
+                  style: Theme.of(context).textTheme.labelMedium),
               Row(
                 children: [
-                  SizedBox(width: 150, child: Text(S.of(context).screenRecipeEditUseSteam)),
+                  SizedBox(
+                      width: 150,
+                      child: Text(S.of(context).screenRecipeEditUseSteam)),
                   ReactiveSwitch(
                     formControlName: 'useSteam',
                     onChanged: (control) => setState(() {}),
@@ -450,7 +478,9 @@ class RecipeEditState extends State<RecipeEdit> {
               const Divider(),
               Row(
                 children: [
-                  SizedBox(width: 150, child: Text(S.of(context).screenRecipeEditUseWater)),
+                  SizedBox(
+                      width: 150,
+                      child: Text(S.of(context).screenRecipeEditUseWater)),
                   ReactiveSwitch(
                     formControlName: 'useWater',
                     onChanged: (control) => setState(() {}),
@@ -465,8 +495,10 @@ class RecipeEditState extends State<RecipeEdit> {
                     keyboardType: const TextInputType.numberWithOptions(),
                     showErrors: (control) => control.invalid,
                     validationMessages: {
-                      ValidationMessage.max: (error) => 'A value greater than ${(error as Map)['max']} is not accepted',
-                      ValidationMessage.min: (error) => 'A value lower than ${(error as Map)['min']} is not accepted',
+                      ValidationMessage.max: (error) =>
+                          'A value greater than ${(error as Map)['max']} is not accepted',
+                      ValidationMessage.min: (error) =>
+                          'A value lower than ${(error as Map)['min']} is not accepted',
                     },
                     decoration: const InputDecoration(
                       labelText: 'Water weight',
@@ -495,11 +527,14 @@ class RecipeEditState extends State<RecipeEdit> {
   void saveFormData(FormGroup form) {
     _editedRecipe.name = form.value["name"] as String;
     _editedRecipe.description = form.value["description"] as String;
-    _editedRecipe.adjustedPressure = form.value["adjustedPressure"] as double? ?? 0;
+    _editedRecipe.adjustedPressure =
+        form.value["adjustedPressure"] as double? ?? 0;
     _editedRecipe.adjustedTemp = form.value["adjustedTemp"] as double? ?? 0;
     _editedRecipe.adjustedWeight = form.value["adjustedWeight"] as double? ?? 0;
-    _editedRecipe.grinderDoseWeight = form.value["grinderDoseWeight"] as double? ?? 0;
-    _editedRecipe.grinderSettings = form.value["grinderSettings"] as double? ?? 0;
+    _editedRecipe.grinderDoseWeight =
+        form.value["grinderDoseWeight"] as double? ?? 0;
+    _editedRecipe.grinderSettings =
+        form.value["grinderSettings"] as double? ?? 0;
     _editedRecipe.grinderModel = form.value["grinderModel"] as String? ?? "";
     _editedRecipe.ratio1 = form.value["ratio1"] as double? ?? 0;
     _editedRecipe.ratio2 = form.value["ratio2"] as double? ?? 0;
@@ -509,7 +544,8 @@ class RecipeEditState extends State<RecipeEdit> {
     _editedRecipe.timeWater = form.value["timeWater"] as double? ?? 0;
     _editedRecipe.useSteam = form.value["useSteam"] as bool;
     _editedRecipe.useWater = form.value["useWater"] as bool;
-    _editedRecipe.disableStopOnWeight = form.value["disableStopOnWeight"] as bool;
+    _editedRecipe.disableStopOnWeight =
+        form.value["disableStopOnWeight"] as bool;
     _editedRecipe.weightMilk = form.value["weightMilk"] as double? ?? 0;
     _editedRecipe.weightWater = form.value["weightWater"] as double? ?? 0;
 
