@@ -159,11 +159,8 @@ class SteamScreenState extends State<SteamScreen> {
                       ],
                     ),
                   ),
-                  const Divider(
-                    height: 20,
-                    thickness: 5,
-                    indent: 20,
-                    endIndent: 0,
+                  Divider(
+                    color: Theme.of(context).colorScheme.primary,
                   ),
                   Padding(
                     padding: const EdgeInsets.all(8.0),
@@ -237,11 +234,8 @@ class SteamScreenState extends State<SteamScreen> {
                         ),
                       ],
                     ),
-                  const Divider(
-                    height: 20,
-                    thickness: 5,
-                    indent: 20,
-                    endIndent: 0,
+                  Divider(
+                    color: Theme.of(context).colorScheme.primary,
                   ),
                   Padding(
                     padding: const EdgeInsets.all(8.0),
@@ -255,10 +249,10 @@ class SteamScreenState extends State<SteamScreen> {
                                   style: Theme.of(context).textTheme.labelLarge),
                               Slider(
                                 value: settings.targetSteamFlow,
-                                max: 4.5,
-                                min: 0.5,
-                                divisions: 40,
-                                label: "${settings.targetSteamFlow.toStringAsFixed(1)} ml/s",
+                                max: 2.5,
+                                min: 0.25,
+                                divisions: 22,
+                                label: "${settings.targetSteamFlow.toStringAsFixed(2)} ml/s",
                                 onChanged: (double value) {
                                   setState(() {
                                     settings.targetSteamFlow = value;
