@@ -9,6 +9,7 @@ import 'package:despresso/model/shotstate.dart';
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
 import 'package:despresso/model/services/state/screen_saver.dart';
+import 'package:flutter_glow/flutter_glow.dart';
 
 import '../../model/services/ble/machine_service.dart';
 import '../../service_locator.dart';
@@ -68,7 +69,8 @@ class _MachineFooterState extends State<MachineFooter> {
   }
 
   bool isOn(EspressoMachineState? state) {
-    return state != EspressoMachineState.sleep && state != EspressoMachineState.disconnected;
+    return state != EspressoMachineState.sleep &&
+        state != EspressoMachineState.disconnected;
   }
 
   @override
@@ -87,11 +89,15 @@ class _MachineFooterState extends State<MachineFooter> {
               builder: (context, snapshot) {
                 return Row(
                   children: snapshot.data != null &&
-                          machineService.currentFullState.state != EspressoMachineState.espresso &&
-                          machineService.currentFullState.state != EspressoMachineState.water &&
-                          machineService.currentFullState.state != EspressoMachineState.steam
+                          machineService.currentFullState.state !=
+                              EspressoMachineState.espresso &&
+                          machineService.currentFullState.state !=
+                              EspressoMachineState.water &&
+                          machineService.currentFullState.state !=
+                              EspressoMachineState.steam
                       ? [
-                          machineService.currentFullState.state == EspressoMachineState.refill
+                          machineService.currentFullState.state ==
+                                  EspressoMachineState.refill
                               ? Container(
                                   color: Colors.red,
                                   child: FooterValue(
@@ -108,42 +114,64 @@ class _MachineFooterState extends State<MachineFooter> {
                 );
               }),
           const Spacer(),
-          if (settingsService.hasSteamThermometer) ThermprobeFooter(machineService: machineService),
+          if (settingsService.hasSteamThermometer)
+            ThermprobeFooter(machineService: machineService),
           StreamBuilder<ShotState>(
               stream: machineService.streamShotState,
               builder: (context, snapshot) {
                 return Row(
-                  children: snapshot.data != null && machineService.currentFullState.state != EspressoMachineState.sleep
+                  children: snapshot.data != null &&
+                          machineService.currentFullState.state !=
+                              EspressoMachineState.sleep
                       ? [
-                          if (settingsService.hasScale) ScaleFooter(machineService: machineService, index: 0),
-                          if (machineService.state.coffeeState != EspressoMachineState.espresso &&
-                              machineService.state.coffeeState != EspressoMachineState.flush &&
-                              machineService.state.coffeeState != EspressoMachineState.steam &&
-                              machineService.state.coffeeState != EspressoMachineState.water &&
+                          if (settingsService.hasScale)
+                            ScaleFooter(
+                                machineService: machineService, index: 0),
+                          if (machineService.state.coffeeState !=
+                                  EspressoMachineState.espresso &&
+                              machineService.state.coffeeState !=
+                                  EspressoMachineState.flush &&
+                              machineService.state.coffeeState !=
+                                  EspressoMachineState.steam &&
+                              machineService.state.coffeeState !=
+                                  EspressoMachineState.water &&
                               settingsService.hasScale &&
                               machineService.scaleService.hasSecondaryScale)
-                            ScaleFooter(machineService: machineService, index: 1),
+                            ScaleFooter(
+                                machineService: machineService, index: 1),
                           FooterValue(
-                              value: "${snapshot.data?.headTemp.toStringAsFixed(1)} °C",
+                              value:
+                                  "${snapshot.data?.headTemp.toStringAsFixed(1)} °C",
                               label: S.of(context).footerGroup),
-                          if (machineService.currentFullState.state != EspressoMachineState.idle)
+                          if (machineService.currentFullState.state !=
+                              EspressoMachineState.idle)
                             FooterValue(
-                                value: "${snapshot.data?.groupPressure.toStringAsFixed(1)} bar",
+                                value:
+                                    "${snapshot.data?.groupPressure.toStringAsFixed(1)} bar",
                                 label: S.of(context).pressure),
-                          if (machineService.currentFullState.state != EspressoMachineState.idle)
+                          if (machineService.currentFullState.state !=
+                              EspressoMachineState.idle)
                             FooterValue(
-                                value: "${snapshot.data?.groupFlow.toStringAsFixed(1)} ml/s",
+                                value:
+                                    "${snapshot.data?.groupFlow.toStringAsFixed(1)} ml/s",
                                 label: S.of(context).flow),
                         ]
                       : [
-                          if (settingsService.hasScale) ScaleFooter(machineService: machineService, index: 0),
-                          if (machineService.state.coffeeState != EspressoMachineState.espresso &&
-                              machineService.state.coffeeState != EspressoMachineState.flush &&
-                              machineService.state.coffeeState != EspressoMachineState.steam &&
-                              machineService.state.coffeeState != EspressoMachineState.water &&
+                          if (settingsService.hasScale)
+                            ScaleFooter(
+                                machineService: machineService, index: 0),
+                          if (machineService.state.coffeeState !=
+                                  EspressoMachineState.espresso &&
+                              machineService.state.coffeeState !=
+                                  EspressoMachineState.flush &&
+                              machineService.state.coffeeState !=
+                                  EspressoMachineState.steam &&
+                              machineService.state.coffeeState !=
+                                  EspressoMachineState.water &&
                               settingsService.hasScale &&
                               machineService.scaleService.hasSecondaryScale)
-                            ScaleFooter(machineService: machineService, index: 1),
+                            ScaleFooter(
+                                machineService: machineService, index: 1),
                         ],
                 );
               }),
@@ -152,26 +180,33 @@ class _MachineFooterState extends State<MachineFooter> {
               stream: machineService.streamState,
               builder: (context, snapshot) {
                 return (snapshot.hasData &&
-                        snapshot.data?.state != EspressoMachineState.disconnected &&
+                        snapshot.data?.state !=
+                            EspressoMachineState.disconnected &&
                         snapshot.data?.state != EspressoMachineState.connecting)
                     ? Row(
                         children: [
-                          Text(isOn(snapshot.data?.state) ? S.of(context).on : S.of(context).off),
-                          Switch(
-                            value: isOn(snapshot.data?.state), //set true to enable switch by default
-                            onChanged: (bool value) {
-                              if (value) {
-                                machineService.de1?.switchOn();
-                              } else {
-                                if (settingsService.screensaverOnIfIdle) {
-                                  var screensaver = getIt<ScreensaverService>();
-                                  screensaver.activateScreenSaver();
-                                }
-                                // screensaver.notifyDelayed();
-                                machineService.de1?.switchOff();
-                              }
-                            },
-                          ),
+                          Padding(
+                              padding: EdgeInsets.only(right: 32),
+                              child: GestureDetector(
+                                  onTap: () {
+                                    isOn(snapshot.data?.state)
+                                        ? machineService.de1?.switchOff()
+                                        : machineService.de1?.switchOn();
+                                  },
+                                  child: isOn(snapshot.data?.state)
+                                      ? GlowIcon(
+                                          Icons.power_settings_new,
+                                          color: Theme.of(context)
+                                              .colorScheme
+                                              .primary,
+                                          glowColor: Theme.of(context)
+                                              .colorScheme
+                                              .primary,
+                                          blurRadius: 9,
+                                          size: 48,
+                                        )
+                                      : GlowIcon(Icons.power_settings_new,
+                                          size: 48)))
                         ],
                       )
                     : const Row();
@@ -210,7 +245,8 @@ class ThermprobeFooter extends StatelessWidget {
                   return Row(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: [
-                      if (machineService.tempService.state == TempState.disconnected)
+                      if (machineService.tempService.state ==
+                          TempState.disconnected)
                         OutlinedButton(
                           onPressed: () {
                             machineService.tempService.connect();
@@ -219,7 +255,8 @@ class ThermprobeFooter extends StatelessWidget {
                             S.of(context).footerConnect,
                           ),
                         ),
-                      if (machineService.tempService.state == TempState.connected)
+                      if (machineService.tempService.state ==
+                          TempState.connected)
                         SizedBox(
                           width: 100,
                           child: Row(
@@ -228,18 +265,24 @@ class ThermprobeFooter extends StatelessWidget {
                                 width: 100,
                                 child: FittedBox(
                                   fit: BoxFit.fitHeight,
-                                  child: machineService.tempService.state == TempState.connected
+                                  child: machineService.tempService.state ==
+                                          TempState.connected
                                       ? Text(
                                           textAlign: TextAlign.right,
                                           "${snapshot.data?.temp1.toStringAsFixed(1)} °C",
-                                          style: Theme.of(context).textTheme.headlineSmall,
+                                          style: Theme.of(context)
+                                              .textTheme
+                                              .headlineSmall,
                                         )
                                       : FittedBox(
                                           fit: BoxFit.fitWidth,
                                           child: Text(
                                             textAlign: TextAlign.right,
-                                            machineService.tempService.state.name,
-                                            style: Theme.of(context).textTheme.headlineSmall,
+                                            machineService
+                                                .tempService.state.name,
+                                            style: Theme.of(context)
+                                                .textTheme
+                                                .headlineSmall,
                                           ),
                                         ),
                                 ),
@@ -273,10 +316,15 @@ class ThermprobeFooter extends StatelessWidget {
           StreamBuilder<Object>(
               stream: machineService.tempService.streamBattery,
               builder: (context, snapshot) {
-                var bat = snapshot.hasData ? (snapshot.data as int) / 100.0 : 0.0;
+                var bat =
+                    snapshot.hasData ? (snapshot.data as int) / 100.0 : 0.0;
                 return LinearProgressIndicator(
                   backgroundColor: Colors.black38,
-                  color: bat < 40 ? Theme.of(context).progressIndicatorTheme.linearTrackColor : Colors.red,
+                  color: bat < 40
+                      ? Theme.of(context)
+                          .progressIndicatorTheme
+                          .linearTrackColor
+                      : Colors.red,
                   value: bat,
                   semanticsLabel: S.of(context).footerBattery,
                 );
@@ -299,28 +347,39 @@ class ScaleFooter extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var showWeightIn = (machineService.state.coffeeState == EspressoMachineState.idle) &&
-        ((machineService.scaleService.hasSecondaryScale && index == 1) ||
-            (machineService.scaleService.hasSecondaryScale == false && index == 0));
+    var showWeightIn =
+        (machineService.state.coffeeState == EspressoMachineState.idle) &&
+            ((machineService.scaleService.hasSecondaryScale && index == 1) ||
+                (machineService.scaleService.hasSecondaryScale == false &&
+                    index == 0));
 
-    var showFlowIn = (machineService.state.coffeeState == EspressoMachineState.espresso) &&
-        machineService.scaleService.hasPrimaryScale &&
-        index == 0;
+    var showFlowIn =
+        (machineService.state.coffeeState == EspressoMachineState.espresso) &&
+            machineService.scaleService.hasPrimaryScale &&
+            index == 0;
     return SizedBox(
       width: (showWeightIn == false && showFlowIn == false) ? 210 : 310,
       child: StreamBuilder<WeightMeassurement>(
-          stream: index == 0 ? machineService.scaleService.stream0 : machineService.scaleService.stream1,
+          stream: index == 0
+              ? machineService.scaleService.stream0
+              : machineService.scaleService.stream1,
           builder: (context, snapshot) {
-            showWeightIn = (machineService.state.coffeeState == EspressoMachineState.idle) &&
-                ((machineService.scaleService.hasSecondaryScale && index == 1) ||
-                    (machineService.scaleService.hasSecondaryScale == false && index == 0));
+            showWeightIn = (machineService.state.coffeeState ==
+                    EspressoMachineState.idle) &&
+                ((machineService.scaleService.hasSecondaryScale &&
+                        index == 1) ||
+                    (machineService.scaleService.hasSecondaryScale == false &&
+                        index == 0));
 
-            showFlowIn = (machineService.state.coffeeState == EspressoMachineState.espresso) &&
+            showFlowIn = (machineService.state.coffeeState ==
+                    EspressoMachineState.espresso) &&
                 machineService.scaleService.hasPrimaryScale &&
                 index == 0;
             return Container(
-              color: machineService.scaleService.state[index] != ScaleState.connecting
-                  ? machineService.scaleService.state[index] != ScaleState.connected
+              color: machineService.scaleService.state[index] !=
+                      ScaleState.connecting
+                  ? machineService.scaleService.state[index] !=
+                          ScaleState.connected
                       ? Colors.red
                       : null
                   : Colors.orange.shade900,
@@ -331,58 +390,88 @@ class ScaleFooter extends StatelessWidget {
                   SizedBox(
                     height: 36,
                     child: StreamBuilder<WeightMeassurement>(
-                        stream: index == 0 ? machineService.scaleService.stream0 : machineService.scaleService.stream1,
+                        stream: index == 0
+                            ? machineService.scaleService.stream0
+                            : machineService.scaleService.stream1,
                         builder: (context, snapshot) {
                           return Row(
                             mainAxisAlignment: MainAxisAlignment.center,
                             children: [
-                              if (machineService.scaleService.state[index] != ScaleState.connecting &&
-                                  machineService.scaleService.state[index] == ScaleState.connected)
+                              if (machineService.scaleService.state[index] !=
+                                      ScaleState.connecting &&
+                                  machineService.scaleService.state[index] ==
+                                      ScaleState.connected)
                                 OutlinedButton(
                                   onPressed: () async {
-                                    await machineService.scaleService.tare(index: index);
+                                    await machineService.scaleService
+                                        .tare(index: index);
                                   },
-                                  child: Text(S.of(context).footerTare),
+                                  child: Text(S.of(context).footerTare,
+                                      style: Theme.of(context)
+                                          .textTheme
+                                          .bodyMedium),
                                 ),
-                              if (machineService.scaleService.state[index] != ScaleState.connecting &&
-                                  machineService.scaleService.state[index] != ScaleState.connected)
+                              if (machineService.scaleService.state[index] !=
+                                      ScaleState.connecting &&
+                                  machineService.scaleService.state[index] !=
+                                      ScaleState.connected)
                                 ElevatedButton(
                                   onPressed: () {
                                     machineService.scaleService.connect();
                                   },
                                   child: Text(
                                     S.of(context).footerConnect,
+                                    style:
+                                        Theme.of(context).textTheme.bodyMedium,
                                   ),
                                 ),
 
-                              if (machineService.scaleService.state[index] == ScaleState.connecting)
+                              if (machineService.scaleService.state[index] ==
+                                  ScaleState.connecting)
                                 Text(
                                   textAlign: TextAlign.right,
                                   machineService.scaleService.state[index].name,
-                                  style: Theme.of(context).textTheme.labelMedium,
+                                  style:
+                                      Theme.of(context).textTheme.labelMedium,
                                 ),
-                              if (machineService.scaleService.state[index] == ScaleState.connected)
+                              if (machineService.scaleService.state[index] ==
+                                  ScaleState.connected)
                                 SizedBox(
-                                  width: (showWeightIn == false && showFlowIn == false) ? 100 : 190,
+                                  width: (showWeightIn == false &&
+                                          showFlowIn == false)
+                                      ? 100
+                                      : 190,
                                   child: Row(
                                     children: [
                                       SizedBox(
                                         width: 100,
                                         child: FittedBox(
                                           fit: BoxFit.fitHeight,
-                                          child: machineService.scaleService.state[index] == ScaleState.connected
+                                          child: machineService.scaleService
+                                                      .state[index] ==
+                                                  ScaleState.connected
                                               ? Text(
                                                   textAlign: TextAlign.right,
                                                   "${snapshot.data?.weight.toStringAsFixed(1)} g",
-                                                  style: Theme.of(context).textTheme.headlineSmall,
+                                                  style: Theme.of(context)
+                                                      .textTheme
+                                                      .headlineSmall,
                                                 )
-                                              : machineService.scaleService.state[index] != ScaleState.disconnected
+                                              : machineService.scaleService
+                                                          .state[index] !=
+                                                      ScaleState.disconnected
                                                   ? FittedBox(
                                                       fit: BoxFit.fitWidth,
                                                       child: Text(
-                                                        textAlign: TextAlign.right,
-                                                        machineService.scaleService.state[index].name,
-                                                        style: Theme.of(context).textTheme.labelSmall,
+                                                        textAlign:
+                                                            TextAlign.right,
+                                                        machineService
+                                                            .scaleService
+                                                            .state[index]
+                                                            .name,
+                                                        style: Theme.of(context)
+                                                            .textTheme
+                                                            .labelSmall,
                                                       ),
                                                     )
                                                   : const Text(""),
@@ -394,18 +483,31 @@ class ScaleFooter extends StatelessWidget {
                                           child: OutlinedButton(
                                             onPressed: () {
                                               Feedback.forTap(context);
-                                              var coffeeService = getIt<CoffeeService>();
-                                              var settingsService = getIt<SettingsService>();
-                                              var r = coffeeService.currentRecipe;
+                                              var coffeeService =
+                                                  getIt<CoffeeService>();
+                                              var settingsService =
+                                                  getIt<SettingsService>();
+                                              var r =
+                                                  coffeeService.currentRecipe;
                                               if (r != null) {
-                                                r.grinderDoseWeight = machineService.scaleService.weight[index];
+                                                r.grinderDoseWeight =
+                                                    machineService.scaleService
+                                                        .weight[index];
                                                 r.adjustedWeight =
-                                                    machineService.scaleService.weight[index] * (r.ratio2 / r.ratio1);
+                                                    machineService.scaleService
+                                                            .weight[index] *
+                                                        (r.ratio2 / r.ratio1);
                                                 coffeeService.updateRecipe(r);
-                                                settingsService.targetEspressoWeight = r.adjustedWeight;
+                                                settingsService
+                                                        .targetEspressoWeight =
+                                                    r.adjustedWeight;
                                               }
                                             },
-                                            child: const FittedBox(child: Text("Set-in")),
+                                            child: FittedBox(
+                                                child: Text("Set-in",
+                                                    style: Theme.of(context)
+                                                        .textTheme
+                                                        .bodyMedium)),
                                           ),
                                         ),
                                       if (showFlowIn)
@@ -413,10 +515,14 @@ class ScaleFooter extends StatelessWidget {
                                           width: 90,
                                           child: Text(
                                             textAlign: TextAlign.right,
-                                            machineService.scaleService.state[index] == ScaleState.connected
+                                            machineService.scaleService
+                                                        .state[index] ==
+                                                    ScaleState.connected
                                                 ? "${snapshot.data?.flow.toStringAsFixed(1)} g/s"
                                                 : "",
-                                            style: Theme.of(context).textTheme.bodyMedium,
+                                            style: Theme.of(context)
+                                                .textTheme
+                                                .bodyMedium,
                                           ),
                                         ),
                                     ],
@@ -435,16 +541,23 @@ class ScaleFooter extends StatelessWidget {
                     S.of(context).footerScale,
                     style: Theme.of(context).textTheme.labelSmall,
                   ),
-                  if (machineService.scaleService.state[index] == ScaleState.connected)
+                  if (machineService.scaleService.state[index] ==
+                      ScaleState.connected)
                     StreamBuilder<Object>(
                         stream: index == 0
                             ? machineService.scaleService.streamBattery0
                             : machineService.scaleService.streamBattery1,
                         builder: (context, snapshot) {
-                          var bat = snapshot.hasData ? ((snapshot.data as BatteryLevel).level) / 100.0 : 0.0;
+                          var bat = snapshot.hasData
+                              ? ((snapshot.data as BatteryLevel).level) / 100.0
+                              : 0.0;
                           return LinearProgressIndicator(
                             backgroundColor: Colors.black38,
-                            color: bat < 40 ? Theme.of(context).progressIndicatorTheme.linearTrackColor : Colors.red,
+                            color: bat < 40
+                                ? Theme.of(context)
+                                    .progressIndicatorTheme
+                                    .linearTrackColor
+                                : Colors.red,
                             value: bat,
                             semanticsLabel: S.of(context).footerBattery,
                           );
@@ -458,7 +571,9 @@ class ScaleFooter extends StatelessWidget {
 }
 
 class FooterValue extends StatelessWidget {
-  const FooterValue({Key? key, required this.value, required this.label, this.width = 120}) : super(key: key);
+  const FooterValue(
+      {Key? key, required this.value, required this.label, this.width = 120})
+      : super(key: key);
 
   final String value;
   final String label;
